refactor: relocate `AresClient` class
