+++ src/modules/events/aresEventManager.ts
@@ -1,7 +1,7 @@
 import { ClientEvents, Collection } from "discord.js";
 import { readdir } from "fs/promises";
 import path from "path";
-import { AresClient } from "../../client";
+import { AresClient } from "../../lib/classes/aresClient";
 import { EventCollection } from "../../ts/types/types";
 import { isEventType } from "../../util/helpers/stringUtil";
 import { LoggerScopes } from "../../util/loggerScopes";
